LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_1-img-src-expected.txt

default-src should cascade to img-src directive


PASS Verify cascading of default-src to img-src policy
FAIL Should fire violation events for every failed violation assert_equals: expected "img-src" but got "default-src"
